% Auto-generated: do not edit by hand
\name{bhBrushableHeatmap}

\alias{bhBrushableHeatmap}

\title{BrushableHeatmap component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
bhBrushableHeatmap(id=NULL, data=NULL, fontSize=NULL, height=NULL,
histLabel=NULL, padding=NULL, redraw=NULL, selectedId=NULL,
selection=NULL, showPoints=NULL, valueLabel=NULL,
width=NULL, xLabel=NULL, yLabel=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Named list. }

\item{fontSize}{Numeric. }

\item{height}{Numeric. }

\item{histLabel}{Character. }

\item{padding}{Numeric. }

\item{redraw}{Logical. }

\item{selectedId}{Numeric. }

\item{selection}{Unnamed list. }

\item{showPoints}{Logical. }

\item{valueLabel}{Character. }

\item{width}{Numeric. Properties for the state of the figure}

\item{xLabel}{Character. }

\item{yLabel}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

